misc/HtmlTestSuite/results/test-0381.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Color Adjustment Test: color-scheme does not affect :root background</title>
<link href="https://drafts.csswg.org/css-color-adjust/#color-scheme-processing" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
  :root { color-scheme: dark }
</style>
</head><body><script>
  test(() =&gt; {
    assert_equals(getComputedStyle(document.documentElement).backgroundColor, "rgba(0, 0, 0, 0)");
  }, "Root element background should be transparent for dark color-scheme.");
</script>
</body>
</html>
